requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.3.4
python-multipart==0.0.9
jinja2==3.1.4
openai==1.35.0
python-dotenv==1.0.1

# Testing
pytest
pytest-asyncio
httpx
